requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
pydantic-settings>=2.2.0
httpx>=0.27.0
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.1.18
chromadb>=0.4.24
python-dotenv>=1.0.1
apscheduler>=3.10.4
python-multipart>=0.0.9
